pmm zone updates

+++ arch/x86_64/arch.c
@@ -1,6 +1,6 @@
 #include <arch/arch.h>
 
-void __attribute__((weak)) arch_map_page(uint64_t va, uint64_t pa, uint64_t flags, uint64_t page_table)
+void __attribute__((weak)) arch_map_page(uint64_t va, uint64_t pa, uint64_t flags, uintptr_t page_table)
 {
     (void) va;
     (void) pa;
@@ -8,7 +8,7 @@ void __attribute__((weak)) arch_map_page(uint64_t va, uint64_t pa, uint64_t flag
     (void) page_table;
 }
 
-void __attribute__((weak)) arch_unmap_page(uint64_t va, uint64_t page_table)
+void __attribute__((weak)) arch_unmap_page(uint64_t va, uintptr_t page_table)
 {
     (void) va;
     (void) page_table;
